refactor(plugin-sdk): group runtime schema boundaries
